extract the new form chart from the technical analysis component
